+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Print of the Python version to the test coverage running
 - Optional dependencies for HTTP/2 and Brotli support in httpx
+- `add_headers` to allow users to update the client's headers
 
 ### Fixed